[issue-158] Moved mlspin information to is't own directory. Added a global pull_homes script
